+++ osquery/events/tests/events_tests.cpp
@@ -110,7 +110,7 @@ TEST_F(EventsTests, test_create_event_pub) {
   EXPECT_TRUE(status.ok());
 
   // Make sure only the first event type was recorded.
-  EXPECT_EQ(EventFactory::numEventPublishers(), 1);
+  EXPECT_EQ(EventFactory::numEventPublishers(), 1U);
 }
 
 class UniqueEventPublisher
@@ -121,11 +121,11 @@ class UniqueEventPublisher
 TEST_F(EventsTests, test_create_using_registry) {
   // The events API uses attachEvents to move registry event publishers and
   // subscribers into the events factory.
-  EXPECT_EQ(EventFactory::numEventPublishers(), 0);
+  EXPECT_EQ(EventFactory::numEventPublishers(), 0U);
   attachEvents();
 
   // Store the number of default event publishers (in core).
-  int default_publisher_count = EventFactory::numEventPublishers();
+  size_t default_publisher_count = EventFactory::numEventPublishers();
 
   // Now add another registry item, but do not yet attach it.
   auto UniqueEventPublisherRegistryItem =
@@ -134,7 +134,7 @@ TEST_F(EventsTests, test_create_using_registry) {
 
   // Now attach and make sure it was added.
   attachEvents();
-  EXPECT_EQ(EventFactory::numEventPublishers(), default_publisher_count + 1);
+  EXPECT_EQ(EventFactory::numEventPublishers(), default_publisher_count + 1U);
 }
 
 TEST_F(EventsTests, test_create_subscription) {
@@ -153,7 +153,7 @@ TEST_F(EventsTests, test_create_subscription) {
   EXPECT_TRUE(status.ok());
 
   // Make sure the subscription is added.
-  EXPECT_EQ(EventFactory::numSubscriptions("publisher"), 1);
+  EXPECT_EQ(EventFactory::numSubscriptions("publisher"), 1U);
 }
 
 TEST_F(EventsTests, test_multiple_subscriptions) {
@@ -166,7 +166,7 @@ TEST_F(EventsTests, test_multiple_subscriptions) {
   status = EventFactory::addSubscription("publisher", subscription);
   status = EventFactory::addSubscription("publisher", subscription);
 
-  EXPECT_EQ(EventFactory::numSubscriptions("publisher"), 2);
+  EXPECT_EQ(EventFactory::numSubscriptions("publisher"), 2U);
 }
 
 struct TestSubscriptionContext : public SubscriptionContext {
@@ -222,7 +222,7 @@ TEST_F(EventsTests, test_create_custom_event_pub) {
 
   // These event types have unique event type IDs
   EXPECT_TRUE(status.ok());
-  EXPECT_EQ(EventFactory::numEventPublishers(), 2);
+  EXPECT_EQ(EventFactory::numEventPublishers(), 2U);
 
   // Make sure the setUp function was called.
   EXPECT_EQ(pub->getTestValue(), 1);
@@ -240,7 +240,7 @@ TEST_F(EventsTests, test_custom_subscription) {
   // Step 3, add the subscription to the event type
   status = EventFactory::addSubscription("TestPublisher", "TestSubscriber", sc);
   EXPECT_TRUE(status.ok());
-  EXPECT_EQ(pub->numSubscriptions(), 1);
+  EXPECT_EQ(pub->numSubscriptions(), 1U);
 
   // The event type must run configure for each added subscription.
   EXPECT_TRUE(pub->configure_run);
@@ -267,7 +267,7 @@ TEST_F(EventsTests, test_tear_down) {
   EXPECT_EQ(pub->getTestValue(), 4);
 
   // Make sure the factory state represented.
-  EXPECT_EQ(EventFactory::numEventPublishers(), 0);
+  EXPECT_EQ(EventFactory::numEventPublishers(), 0U);
 }
 
 static int kBellHathTolled = 0;
@@ -334,7 +334,7 @@ TEST_F(EventsTests, test_event_sub_subscribe) {
 
   // Don't overload the normal `init` Subscription member.
   sub->lateInit();
-  EXPECT_EQ(pub->numSubscriptions(), 1);
+  EXPECT_EQ(pub->numSubscriptions(), 1U);
 
   auto ec = pub->createEventContext();
   pub->fire(ec, 0);
